// packages/plugin/colors/types.ts
export type ColorScale =
    | Partial<{
          50: string;
          100: string;
          200: string;
          300: string;
          400: string;
          500: string;
          600: string;
          700: string;
          800: string;
          900: string;
          foreground: string;
          DEFAULT: string;
      }>
    | string;

export type BaseColors = {
    background: ColorScale;
    foreground: ColorScale;
    content1: ColorScale;
    content2: ColorScale;
    content3: ColorScale;
    content4: ColorScale;
};

export type ThemeColors = BaseColors & {
    default: ColorScale;
    primary: ColorScale;
    secondary: ColorScale;
    accent: ColorScale;
    success: ColorScale;
    warning: ColorScale;
    danger: ColorScale;
};

export type SemanticBaseColors = {
    light: BaseColors;
    dark: BaseColors;
};